requirements.txt
kubernetes==11.0.0
pytest==5.4.3
cryptography==2.9.2
google-cloud-secret-manager==0.2.0